requirements.txt
transformers>=4.44.2
torch>=2.6.0
pandas>=2.1.0
numpy>=1.24.0,<2.0.0
yfinance>=0.2.33
scikit-learn==1.6.1
joblib>=1.3.0
pyyaml>=6.0.0
huggingface_hub>=0.23.0
requests>=2.31.0
omegaconf>=2.3.0
pyarrow>=14.0.0
feedparser>=6.0.11
streamlit>=1.38.0
ruamel.yaml>=0.18.0

xgboost>=2.1.1
